pyproject.toml
[tool.poetry]
name = "gtfs-general"
version = "0.1.0"
description = "A general to command your GTFS data"
authors = ["Julian Psotta <julianpsotta@gmail.com>"]
readme = "README.md"
packages = [{ include = "gtfs_general", from = "src" }]

[tool.poetry.scripts]
gtfs-general = "gtfs_general.main:app"

[tool.poetry.dependencies]
python = ">=3.8,<4"
dask = { extras = ['distributed'], version = "^2022.9.2" }
pandas = "^1.5.0"
typer = { extras = ["all"], version = "^0.6.1" }
rich = "^12.6.0"
tqdm = "^4.64.1"
click = "^8.1.3"
click-extra = "^3.2.5"
uvicorn = { extras = ["standard"], version = "^0.19.0" }
fastapi = "^0.85.1"
gunicorn = "^20.1.0"
geoalchemy2 = "^0.12.5"
tox = "^3.27.0"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.5"
sqlalchemy-stubs = "^0.4"



[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.982"
pytest = "^7.1.3"
bokeh = "^2.4.3"
tox-docker = "^4.0.0a2"
tox = "^3.27.0"
pytest-cov = "^4.0.0"
autoflake = "^1.7.7"
types-PyYAML = "^6.0.12"
pytest-xdist = "^3.0.2"
pydantic = "^1.10.6"

[build-system]
requires = ["poetry-core", "poetry>=1.2"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 121
skip_glob = ["**/.env*", "**/.venv*", "**/env/*", "**/.venv/*", "**/docs/*"]

[tool.black]
line-length = 121
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.coverage
  | \.tox
  | \.venv
  | \.env
  | \.venv310
  | \.venv309
  | \.venv308
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
log_cli = false
log_cli_level = 20 # Warnings
filterwarnings = [
    "ignore:::.*tests.gtfs_general.db.test_gtfs_agency*",
    "ignore:::.*tests.gtfs_general.db.test_gtfs_calendar*",
    "ignore:::.*tests.gtfs_general.db.test_gtfs_routes*",
    "ignore:::.*tests.gtfs_general.db.test_gtfs_stops*",
    "ignore:::.*tests.gtfs_general.db.test_gtfs_shapes*",
    "ignore:::.*tests.gtfs_general.db.test_gtfs_trips*",
]

[tool.mypy]
plugins = [
  #"pydantic.mypy"
]

ignore_missing_imports = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
exclude = '.git,__pycache__,__init__.py,.mypy_cache,.pytest_cache,alembic,.venv,.tox,.idea,.venv310,.env,input,output'

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
